# artrack/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/artrack/tracks/": {
      "post": {
        "tags": [
          "artrack-tracks"
        ],
        "summary": "Create Track",
        "description": "Create a new track",
        "operationId": "create_track_artrack_tracks__post",
        "parameters": [
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "artrack-tracks"
        ],
        "summary": "List Tracks",
        "description": "List tracks",
        "operationId": "list_tracks_artrack_tracks__get",
        "parameters": [
          {
            "name": "visibility",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Visibility"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackResponse"
                  },
                  "title": "Response List Tracks Artrack Tracks  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artrack/tracks/{track_id}": {
      "get": {
        "tags": [
          "artrack-tracks"
        ],
        "summary": "Get Track",
        "description": "Get track details",
        "operationId": "get_track_artrack_tracks__track_id__get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Track Id"
            }
          },
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "artrack-tracks"
        ],
        "summary": "Delete Track",
        "description": "Delete a track",
        "operationId": "delete_track_artrack_tracks__track_id__delete",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Track Id"
            }
          },
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "artrack-tracks"
        ],
        "summary": "Update Track",
        "description": "Update a track",
        "operationId": "update_track_artrack_tracks__track_id__put",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Track Id"
            }
          },
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/artrack/admin/tracks": {
      "get": {
        "tags": [
          "artrack-admin"
        ],
        "summary": "Admin Tracks",
        "operationId": "admin_tracks_artrack_admin_tracks_get",
        "parameters": [
          {
            "name": "visibility",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Visibility"
            }
          },
          {
            "name": "track_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Track Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTracksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/artrack/admin/tracks/{track_id}": {
      "delete": {
        "tags": [
          "artrack-admin"
        ],
        "summary": "Admin Delete Track",
        "description": "Delete a track as admin (bypasses ownership).",
        "operationId": "admin_delete_track_artrack_admin_tracks__track_id__delete",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Track Id"
            }
          },
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminTrackItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility"
          },
          "track_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "waypoint_count": {
            "type": "integer",
            "title": "Waypoint Count"
          },
          "collaborator_count": {
            "type": "integer",
            "title": "Collaborator Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "owner_id",
          "visibility",
          "created_at",
          "waypoint_count",
          "collaborator_count"
        ],
        "title": "AdminTrackItem"
      },
      "AdminTracksResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminTrackItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "AdminTracksResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "TrackCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility",
            "default": "private"
          },
          "track_type": {
            "type": "string",
            "title": "Track Type",
            "default": "hiking"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "client_track_id": {
            "type": "string",
            "title": "Client Track Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "client_track_id"
        ],
        "title": "TrackCreate"
      },
      "TrackResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility"
          },
          "track_type": {
            "type": "string",
            "title": "Track Type"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "client_track_id": {
            "type": "string",
            "title": "Client Track Id"
          },
          "stats": {
            "$ref": "#/components/schemas/TrackStats"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "visibility",
          "track_type",
          "tags",
          "client_track_id",
          "stats",
          "created_at",
          "updated_at"
        ],
        "title": "TrackResponse"
      },
      "TrackStats": {
        "properties": {
          "total_waypoints": {
            "type": "integer",
            "title": "Total Waypoints"
          },
          "processed_waypoints": {
            "type": "integer",
            "title": "Processed Waypoints"
          },
          "pending_analysis": {
            "type": "integer",
            "title": "Pending Analysis"
          },
          "distance_meters": {
            "type": "number",
            "title": "Distance Meters"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "total_waypoints",
          "processed_waypoints",
          "pending_analysis",
          "distance_meters",
          "duration_seconds"
        ],
        "title": "TrackStats"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Visibility": {
        "type": "string",
        "enum": [
          "public",
          "followers",
          "private"
        ],
        "title": "Visibility"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    }
  }
}